requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
requests
pandas
numpy
matplotlib
Pillow
duckdb
lxml
pyarrow